Move files related to render into dedicated folder
