+++ lib/renderer/qrcode-draw.js
@@ -15,7 +15,7 @@
 *
 */
 
-var QRCodeLib = require('./core/qrcode.js')
+var QRCodeLib = require('../core/qrcode.js')
 
 exports.QRCodeDraw = QRCodeDraw
 exports.QRErrorCorrectLevel = QRCodeLib.QRErrorCorrectLevel
