+++ lib/renderer/svg.js
@@ -1,6 +1,4 @@
-module.exports.renderBits = renderBits
-
-function renderBits (bits, width, options) {
+exports.renderBits = function renderBits (bits, width, options) {
   if (typeof bits === 'undefined' || !(bits instanceof Array)) {
     throw new Error('"bits" must be a valid Array')
   }
