+++ lib/renderer/terminal.js
